+++ ntropy_sdk/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "4.9.1rc1"
+__version__ = "4.10.2rc1"
 
 from ntropy_sdk.ntropy_sdk import (
     AccountHolder,