+++ ntropy_sdk/ntropy_sdk.py
@@ -10,9 +10,9 @@ from typing import Any, ClassVar, Generator, List, Optional, TypeVar, Iterable
 from urllib.parse import urlencode
 from itertools import islice
 
-import requests
+import requests  # type: ignore
 from pydantic import BaseModel, Field, validator, NonNegativeFloat
-from requests_toolbelt.adapters.socket_options import TCPKeepAliveAdapter
+from requests_toolbelt.adapters.socket_options import TCPKeepAliveAdapter  # type: ignore
 from tqdm.auto import tqdm
 
 from ntropy_sdk import __version__
@@ -362,6 +362,23 @@ class AccountHolder(BaseModel):
         extra = "allow"
 
 
+class RecurrenceGroup(BaseModel):
+    """Information regarding the recurrence group of one transaction"""
+
+    date_of_first_tx: Optional[date]
+    date_of_last_tx: Optional[date]
+    frequency_in_days: Optional[int]
+    average_amount_per_tx: Optional[float]
+    other_party: Optional[str]
+    id: Optional[str]
+    transaction_ids: Optional[List[str]]
+
+    class Config:
+        use_enum_values = True
+        arbitrary_types_allowed = True
+        extra = "allow"
+
+
 class EnrichedTransaction(BaseModel):
     """An enriched financial transaction."""
 
@@ -378,6 +395,7 @@ class EnrichedTransaction(BaseModel):
         "website",
         "chart_of_accounts",
         "recurrence",
+        "recurrence_group",
         "confidence",
         "transaction_type",
         "mcc",
@@ -398,7 +416,10 @@ class EnrichedTransaction(BaseModel):
         description="Label from the standard chart-of-accounts hierarchy."
     )
     recurrence: Optional[RecurrenceType] = Field(
-        description="Indicates if the Transaction is recurring."
+        description="Indicates if the Transaction is recurring and the type of recurrence"
+    )
+    recurrence_group: Optional[RecurrenceGroup] = Field(
+        description="Contains the information of the recurrence group if the transaction is recurrent"
     )
     confidence: Optional[NonNegativeFloat] = Field(
         description="A numerical score between 0.0 and 1.0 indicating the confidence",
@@ -425,6 +446,11 @@ class EnrichedTransaction(BaseModel):
     def __init__(self, **kwargs):
         fields = {}
         extra = {}
+
+        recurrence_group = self._parse_recurrence_group(kwargs)
+        if recurrence_group is not None:
+            fields["recurrence_group"] = recurrence_group
+
         for key in kwargs:
             if key in EnrichedTransaction._fields:
                 fields[key] = kwargs[key]
@@ -433,6 +459,38 @@ class EnrichedTransaction(BaseModel):
 
         super().__init__(**fields)
         self.kwargs = extra
+
+    def _parse_recurrence_group(self, kwargs: dict) -> Optional[RecurrenceGroup]:
+        def _from_recurrence_v1(kwargs):
+            return RecurrenceGroup(
+                id=kwargs["recurrence_group_id"],
+                transaction_ids=[
+                    x["transaction_id"] for x in kwargs["recurrence_group"]
+                ],
+            )
+
+        def _from_recurrence_v2(kwargs):
+            return RecurrenceGroup(**kwargs["recurrence_group"])
+
+        recurrence_group: Optional[RecurrenceGroup] = None
+        # parse recurrence api v1
+        if all(
+            [
+                x in kwargs
+                for x in ["recurrence", "recurrence_group", "recurrence_group_id"]
+            ]
+        ) and isinstance(kwargs["recurrence_group"], list):
+            recurrence_group = _from_recurrence_v1(kwargs)
+            del kwargs["recurrence_group"]
+            del kwargs["recurrence_group_id"]
+        # parse recurrence api v2
+        elif all(
+            [x in kwargs for x in ["recurrence", "recurrence_group"]]
+        ) and isinstance(kwargs["recurrence_group"], dict):
+            recurrence_group = _from_recurrence_v2(kwargs)
+            del kwargs["recurrence_group"]
+
+        return recurrence_group
 
     def __repr__(self):
         return f"EnrichedTransaction({dict_to_str(self.to_dict())})"
@@ -1287,6 +1345,10 @@ class SDK:
                 error = e.response.json()
                 raise ValueError(f"{error['detail']}")
             raise
+        except AttributeError:
+            raise TypeError(
+                "transactions must be either a pandas.Dataframe or an iterable"
+            )
 
     @singledispatchmethod
     def add_transactions_async(
@@ -1527,7 +1589,7 @@ class SDK:
         response = self.retry_ratelimited_request("POST", url, payload)
         return response.json()
 
-    def get_account_income_report(self, account_holder_id: str) -> dict:
+    def get_account_income_report(self, account_holder_id: str) -> IncomeReport:
         """Returns the income report of an account holder's Transaction history
 
         Parameters
@@ -1547,7 +1609,7 @@ class SDK:
         url = f"/v2/account-holder/{account_holder_id}/income"
 
         response = self.retry_ratelimited_request("POST", url, {})
-        return IncomeReport(response.json())
+        return IncomeReport.from_dicts(response.json())
 
     def get_account_subscription_report(self, account_holder_id: str) -> dict:
         """Returns the subscription report of an account holder's Transaction history